U8, U16, U32 data type changes

+++ src/5gnrsch/sch_utils.h
@@ -47,7 +47,7 @@
 /* allocate and zero out a static buffer */
 #define SCH_ALLOC(_datPtr, _size)                               \
 {                                                               \
-   int _ret;                                                    \
+   uint8_t _ret;                                                    \
    _ret = SGetSBuf(SCH_MEM_REGION, SCH_POOL,                    \
           (Data **)&_datPtr, _size);                             \
    if(_ret == ROK)                                              \
